// src/app/enums/v-card-type3-phone.enum.ts
enum VCardType3PhoneEnum {
  HOME = 'HOME', // to indicate a telephone number associated with a residence
  // MSG = 'MSG', // to indicate the telephone number has voice messaging
  // support
  WORK = 'WORK', // to indicate a telephone number associated with a place of
                 // work
  PREF = 'PREF', // to indicate a preferred-use telephone number
  VOICE = 'VOICE', // to indicate a voice telephone number
  // FAX = 'FAX', // to indicate a facsimile telephone number
  CELL = 'CELL', // to indicate a cellular telephone number
  // VIDEO = 'VIDEO', // to indicate a video conferencing telephone number
  // PAGER = 'PAGER', // to indicate a paging device telephone number
  // BBS = 'BBS', // to indicate a bulletin board system telephone number
  // MODEM = 'MODEM', // to indicate a MODEM connected telephone number
  // CAR = 'CAR', // to indicate a car-phone telephone number
  // ISDN = 'ISDN', // to indicate an ISDN service telephone number
  // PCS = 'PCS', // to indicate a personal communication services telephone
  // number
}

export default VCardType3PhoneEnum